Separate topic_streams module
